Move Particular Assignment into different package. Remove unused const PI Check naming conventions for packages and variables. Add validation in both Problem.
